requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
opencv-python>=4.5.3
streamlit>=1.10.0
ffmpeg-python>=0.2.0
scipy>=1.7.1
matplotlib>=3.4.3
pandas>=1.3.3 